refact: ai 관련 메소드 도메인 이동 ref:#42
